dls_focusstack gui running on Linux
