+++ dls_focusstack/focus/focus_stack_lap_pyramid.py
@@ -6,20 +6,11 @@ import numpy as np
 
 from config.focus_config import FocusConfig
 from dls_util.imaging import Image
+from focus.image_fft_manager import ImageFFTManager
 from focus.pyramid import pyramid
 from os.path import join
 
-from focus.image_fft import Image_FFT
-
-IMG_TO_STACK = 8 #how many images will be stacked
-
-
-def f(file_obj,q,count):
-    img_color = cv2.imread(file_obj.name)
-    img = cv2.cvtColor(img_color.astype(np.float32), cv2.COLOR_BGR2GRAY)
-    image_fft = Image_FFT(img, count)
-    image_fft.runFFT()
-    q.put(image_fft)
+from focus.sharpness_detector import SharpnessDetector
 
 
 class FocusStack:
@@ -31,7 +22,12 @@ class FocusStack:
 
     def composite(self):
         t1 = time.clock()
-        images = self.find_sharp()
+        man = ImageFFTManager(self._image_file_list)
+        man.read_ftt_images()
+        sd = SharpnessDetector(man.get_fft_images())
+
+        images = sd.images_to_stack()
+
         t2 = time.clock() - t1
         print 'time fft:', t2
         images = np.array(images, dtype=images[0].dtype)
@@ -44,58 +40,6 @@ class FocusStack:
         stacked_image  = cv2.convertScaleAbs(stacked_image)
         return Image(stacked_image)
 
-    def find_sharp(self):
-
-        q = Queue()
-
-        #processes = [Process(target=f, args=(file_obj,q)) for file_obj in self._image_file_list]
-        processes=[]
-        count = 1
-        for file_obj in self._image_file_list:
-            process = Process(target=f, args=(file_obj,q,count))
-            processes.append(process)
-            count = count+1
-
-        for p in processes:
-            p.start()
-
-        sd = [q.get() for p in processes]
-
-        for p in processes:
-            p.join()
-
-        images = self.images_to_stack(sd)
-        #take n images from the stuck
-        return images
-
-
-    def images_to_stack(self, sd):
-        n = len(sd)
-        level = 0
-        max = None
-        images = []
-        for s in sd:
-            fft = s.getFFT()
-            if fft > level:
-                level = fft
-                max = s.getImageNumber()
-
-        range = self.find_range(max,n)
-        for s in sd:
-            if s.getImageNumber() in range:
-                images.append(s.getImage())
-
-        return images
-
-
-    @staticmethod
-    def find_range(max,n):
-        if max -(IMG_TO_STACK / 2) < 1:
-            return range(1, IMG_TO_STACK)
-        elif max + (IMG_TO_STACK / 2) > n:
-            return range(-IMG_TO_STACK, n)
-        else:
-            return range(max - IMG_TO_STACK / 2, max + IMG_TO_STACK / 2)
